requirements.txt
fastapi==0.123.0
uvicorn==0.38.0
scikit-learn==1.7.2
scipy==1.16.3
numpy==2.3.5
joblib==1.5.2
pandas==2.2.3
matplotlib==3.9.2
openai-agents==0.6.1
python-dotenv==1.0.1
tqdm==4.66.5
pytest==8.3.3
